feat: better rate limiting, break api into its own app
