+++ package.json
@@ -32,6 +32,7 @@
     "express": "^5.1.0",
     "express-openapi": "^12.1.3",
     "express-rate-limit": "^7.1.4",
+    "express-slow-down": "^3.0.0",
     "file-type": "16",
     "filenamify": "^4.3.0",
     "fluent-ffmpeg": "^2.1.2",
