+++ src/routes.ts
@@ -1,0 +1,130 @@
+// Just a list of all routes in the API.
+// API docs needs this to mount them
+const routes = [
+  "tags",
+  "trackGroups",
+  "trackGroups/topSold",
+  "trackGroups/mostPlayed",
+  "trackGroups/testExistence",
+  "trackGroups/{id}",
+  "trackGroups/{id}/download",
+  "trackGroups/{id}/redeemCode",
+  "trackGroups/{id}/purchase",
+  "trackGroups/{id}/testOwns",
+  "trackGroups/{id}/wishlist",
+  "trackGroups/{id}/emailDownload",
+  "trackGroups/{id}/emailPurchaseLink",
+  "trackGroups/{id}/generate",
+  "playable",
+  "settings/{setting}",
+  "tracks",
+  "tracks/topSold",
+  "tracks/mostPlayed",
+  "tracks/{id}",
+  "tracks/{id}/audio",
+  "tracks/{id}/purchase",
+  "tracks/{id}/download",
+  "tracks/{id}/generate",
+  "tracks/{id}/testOwns",
+  "tracks/{id}/favorite",
+  "tracks/{id}/stream/{segment}",
+  "artists",
+  "artists/testExistence",
+  "artists/{id}",
+  "artists/{id}/subscribe",
+  "artists/{id}/purchaseCatalogue",
+  "artists/{id}/feed",
+  "artists/{id}/follow",
+  "artists/{id}/posts",
+  "artists/{id}/followers",
+  "artists/{id}/supporters",
+  "artists/{id}/inbox",
+  "artists/{id}/confirmFollow",
+  "artists/{id}/unfollow",
+  "artists/{id}/tip",
+  "merch/{id}",
+  "merch/{id}/purchase",
+  "posts",
+  "posts/{id}",
+  "licenses",
+  "labels",
+  "labels/{id}",
+  "labels/{id}/trackGroups",
+  "users",
+  "users/testExistence",
+  "users/{userId}",
+  "users/{userId}/confirmEmail",
+  "users/{userId}/avatar",
+  "users/{userId}/banner",
+  "users/{userId}/notifications",
+  "users/{userId}/notifications/unreadCount",
+  "users/{userId}/notifications/{notificationId}",
+  "users/{userId}/purchases",
+  "users/{userId}/wishlist",
+  "users/{userId}/charges",
+  "users/{userId}/stripe/connect",
+  "users/{userId}/stripe/connectComplete",
+  "users/{userId}/stripe/checkAccountStatus",
+  "users/{userId}/feed",
+  "manage/subscriptions",
+  "manage/subscriptions/{subscriptionId}",
+  "manage/artists",
+  "manage/label",
+  "manage/label/artists/{artistId}",
+  "manage/artists/{artistId}",
+  "manage/artists/{artistId}/trackGroups",
+  "manage/artists/{artistId}/merch",
+  "manage/artists/{artistId}/labels",
+  "manage/artists/{artistId}/labels/{labelUserId}",
+  "manage/artists/{artistId}/codes",
+  "manage/artists/{artistId}/subscribers",
+  "manage/artists/{artistId}/banner",
+  "manage/artists/{artistId}/avatar",
+  "manage/artists/{artistId}/drafts",
+  "manage/artists/{artistId}/subscriptionTiers",
+  "manage/artists/{artistId}/subscriptionTiers/{subscriptionTierId}",
+  "manage/trackGroups/{trackGroupId}",
+  "manage/trackGroups/{trackGroupId}/trackOrder",
+  "manage/trackGroups/{trackGroupId}/publish",
+  "manage/trackGroups/{trackGroupId}/tags",
+  "manage/trackGroups/{trackGroupId}/cover",
+  "manage/trackGroups/{trackGroupId}/codes",
+  "manage/merch/{merchId}",
+  "manage/merch/{merchId}/image",
+  "manage/merch/{merchId}/destinations",
+  "manage/merch/{merchId}/optionTypes",
+  "manage/purchases",
+  "manage/purchases/{purchaseId}",
+  "manage/purchases/{purchaseId}/contactArtist",
+  "manage/sales",
+  "manage/tracks",
+  "manage/tracks/{trackId}/audio",
+  "manage/tracks/{trackId}/downloadOriginal",
+  "manage/tracks/{trackId}",
+  "manage/tracks/{trackId}/trackArtists",
+  "manage/posts/{postId}/images",
+  "manage/posts/{postId}/tracks",
+  "manage/posts/{postId}/featuredImage",
+  "manage/posts/{postId}/publish",
+  "manage/posts/{postId}",
+  "manage/posts",
+  "checkout",
+  "checkout/status",
+  "webhooks/stripe",
+  "webhooks/stripe/connect",
+  "jobs",
+  "admin/tasks",
+  "admin/tracks",
+  "admin/trackGroups",
+  "admin/artists",
+  "admin/subscriptions",
+  "admin/purchases",
+  "admin/settings",
+  "admin/tips",
+  "admin/users",
+  "admin/send-email",
+  "oembed",
+  "flag",
+];
+
+export default routes;
